refactor: refactor shared out completely
