+++ backend/Dockerfile
@@ -4,14 +4,10 @@ WORKDIR /app
 
 # Copy the requirements file into the container
 COPY backend/requirements.txt ./backend/requirements.txt
-COPY shared/requirements.txt ./shared/requirements.txt
 
 # Install the dependencies
 RUN pip install --no-cache-dir -r backend/requirements.txt
-RUN pip install --no-cache-dir -r shared/requirements.txt
 
-# Copy the shared directory
-COPY ./shared ./shared
 # Copy the backend directory
 COPY ./backend ./backend
