--- backend/setup.py
@@ -1,11 +1,0 @@
-# setup.py
-from setuptools import find_packages, setup
-
-setup(
-    name="shared",
-    version="0.1",
-    packages=find_packages(),
-    install_requires=[
-        "pydantic",
-    ],
-)
